refactor: fix misleading names and a duplicate upload call (part of #3491)
Names that say the wrong thing invite confident wrong edits, especially
from automated refactors and agents acting on the name alone:

- prose-guardian-settings.ts: remove shouldHoldForProseGuardianRewrite,
  a verbatim alias of shouldHoldForTextRewrite. The name implied a
  prose-guardian-only gate, but the predicate fires for prose-guardian,
  continuity, AND html rewrites and its result gates seven generation
  branches. generate.routes.ts now calls the real name and the local
  holdForTextRewrite mirrors it.
- GameCombatUI/GameSurface: rename the _isStreaming prop to isStreaming.
  The underscore convention marks unused values, but this prop is
  load-bearing (it drives the custom-action reset effect); an automated
  "remove unused underscore props" pass would have broken combat UI.
- PersonaTrackerPanel.tsx -> PersonaInventoryPanel.tsx: the filename now
  matches its export (PersonaInventoryPanel), restoring the sibling
  file-per-section convention.
- video-generation.ts: the Seedance reference resolver called
  maybeUploadSeedanceReferenceImage twice with identical arguments on
  the raw-URL path (inner call, then the shared fallback). Both paths
  now converge on the single fallback call, which also stops the
  raw-URL path from attempting the same upload twice.

Two #3491 items were re-verified and NOT changed here: the
HomeProfessorMariChat textarea refs are no longer inverted on current
staging (the component was restructured upstream; verified the floating
ref now attaches inside the floating surface), and the retry-agents
chatMeta shadow is left alone (four parallel local parses, no active
bug, re-scoping a mega-route without runtime coverage is riskier than
the latent issue). The remaining #3491 dedups (Mari chat body, sprite
crop helpers) still need behavior-equivalence verification and stay
open on the issue.

Part of #3491
