+++ packages/client/src/features/tracker-panel/components/sections/PersonaInventoryPanel.tsx
